pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-frontmatter"
dynamic = ["version"]
description = 'Manipulate YAML front matter.'
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = []
authors = [
  { name = "YEUNG King On", email = "koyeung@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "ruamel.yaml>=0.17.21",
  "jsonpath-ng>=1.5.3",
]

[project.urls]
Documentation = "https://github.com/koyeung/py-frontmatter#readme"
Issues = "https://github.com/koyeung/py-frontmatter/issues"
Source = "https://github.com/koyeung/py-frontmatter"

[project.scripts]
frontmatter = "py_frontmatter.console.application:main"

[tool.hatch.version]
path = "src/py_frontmatter/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "pytest-cov",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.10", "3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=23.1.0",
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/py_frontmatter tests}"
style = [
  "ruff {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff --fix {args:.}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.coverage.run]
source_pkgs = ["py_frontmatter", "tests"]
branch = true
parallel = true
omit = [
  "src/py_frontmatter/__about__.py",
]

[tool.coverage.paths]
py_frontmatter = ["src/py_frontmatter", "*/py-frontmatter/src/py_frontmatter"]
tests = ["tests", "*/py-frontmatter/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--doctest-modules --no-cov-on-fail --cov-report term-missing:skip-covered"

[tool.ruff]
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.isort]
known-first-party = ["py_frontmatter"]

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
